Partially modularize models
