+++ webpack.config.js
@@ -1,0 +1,7 @@
+module.exports = {
+  entry: "./src/sa_web/static/js/views/app-view.js",
+  output: {
+    path: "./src/sa_web/static/",
+    filename: "test-bundle.js"
+  }
+}
